% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redshift.R
\name{checkAwsS3Connection}
\alias{checkAwsS3Connection}
\title{Check to see if a connection to Amazon S3 is valid and usable}
\usage{
checkAwsS3Connection()
}
\value{
TRUE if the Aws S3 connection is valid and usable, FALSE if not
}
\description{
For Amazon Redshift database connections, this function will check to see if a connection to 
Amazon S3 is configured properly and usable. A valid AWS S3 connection is necessary to 
bulk load data to Redshift when using the \code{\link{insertTable}} function.
}
\details{
The function relies upon the CloudyR S3 library to test a connection to an S3 bucket using
AWS S3 credentials. Credentials are configured either directly into the System Environment
using the following keys: 

Sys.setenv("AWS_ACCESS_KEY_ID" = "some_access_key_id",
           "AWS_SECRET_ACCESS_KEY" = "some_secret_access_key",
           "AWS_DEFAULT_REGION" = "some_aws_region",
           "AWS_BUCKET_NAME" = "some_bucket_name",
           "AWS_OBJECT_KEY" = "some_object_key",
           "AWS_SSE_TYPE" = "server_side_encryption_type")

Alternatively, an AWS credentials file, saved locally in ~./aws/credentials file, can 
auto-populate these System Environment variables if it includes the following keys 
in the default profile:

[default]
aws_access_key_id = some_access_key_id
aws_secret_access_key = some_secret_access_key
aws_default_region = some_aws_region
aws_bucket_name = some_bucket_name
aws_object_key = some_object_key
aws_sse_type = server_side_encryption_type
}
\references{
\url{https://github.com/cloudyr/aws.s3}

\url{https://aws.amazon.com/blogs/security/a-new-and-standardized-way-to-manage-credentials-in-the-aws-sdks/}
}
