% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compression.R
\name{createZipFile}
\alias{createZipFile}
\title{Compress files and/or folders into a single zip file}
\usage{
createZipFile(zipFile, files, rootFolder = getwd(), compressionLevel = 9)
}
\arguments{
\item{zipFile}{The path to the zip file to be created.}

\item{files}{The files and/or folders to be included in the zip file. Folders will be included recursively.}

\item{rootFolder}{The root folder. All files will be stored with relative paths relative to this folder.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the longest.}
}
\description{
Compress files and/or folders into a single zip file
}
\details{
Uses Java's compression library to create a zip file. It is similar to \code{utils::zip}, except
that it does not require an external zip tool to be available on the system path.
}
