% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnector.R
\docType{package}
\name{DatabaseConnector-package}
\alias{DatabaseConnector}
\alias{DatabaseConnector-package}
\title{DatabaseConnector: Connect to OHDSI Supported Database Platforms with DBI}
\description{
An R Database Interface (DBI) compatible interface to several database platforms using Java Database Connectivity (JDBC). Supported platforms are PostgreSQL, Oracle, Microsoft SQL Server, Amazon Redshift, Google BigQuery, Snowflake, Spark.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DatabaseConnector/}
  \item \url{https://github.com/OHDSI/DatabaseConnector}
  \item Report bugs at \url{https://github.com/OHDSI/DatabaseConnector/issues}
}

}
\author{
\strong{Maintainer}: Adam Black \email{black@ohdsi.org} (\href{https://orcid.org/0000-0001-5576-8701}{ORCID})

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Marc Suchard
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
  \item Microsoft Inc. (SQL Server JDBC driver) [copyright holder]
  \item PostgreSQL Global Development Group (PostgreSQL JDBC driver) [copyright holder]
  \item Oracle Inc. (Oracle JDBC driver) [copyright holder]
  \item Amazon Inc. (RedShift JDBC driver) [copyright holder]
}

}
\keyword{internal}
