% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorDbiConnection-class}
\alias{DatabaseConnectorDbiConnection-class}
\title{DatabaseConnectorDbiConnection class.}
\description{
DatabaseConnectorDbiConnection class.
}
\keyword{internal}
