% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorDriver-class}
\alias{DatabaseConnectorDriver-class}
\alias{dbUnloadDriver,DatabaseConnectorDriver-method}
\title{DatabaseConnectorDriver class.}
\usage{
\S4method{dbUnloadDriver}{DatabaseConnectorDriver}(drv, ...)
}
\description{
DatabaseConnectorDriver class.
}
\keyword{internal}

