% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{DatabaseConnectorDriver}
\alias{DatabaseConnectorDriver}
\alias{dbConnect,DatabaseConnectorDriver-method}
\alias{dbDisconnect,DatabaseConnectorConnection-method}
\alias{show,DatabaseConnectorConnection-method}
\title{Create a DatabaseConnectorDriver object}
\usage{
DatabaseConnectorDriver()

\S4method{dbConnect}{DatabaseConnectorDriver}(drv, ...)

\S4method{dbDisconnect}{DatabaseConnectorConnection}(conn)

\S4method{show}{DatabaseConnectorConnection}(object)
}
\arguments{
\item{drv}{An object created by \code{DatabaseConnectorDriver()}}

\item{conn}{An object created by \code{connect()}}

\item{object}{An object created by \code{connect()}}
}
\description{
Create a DatabaseConnectorDriver object
}

