% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorJdbcConnection-class}
\alias{DatabaseConnectorJdbcConnection-class}
\alias{dbIsValid,DatabaseConnectorJdbcConnection-method}
\alias{dbSendQuery,DatabaseConnectorJdbcConnection,character-method}
\title{DatabaseConnectorJdbcConnection class.}
\usage{
\S4method{dbIsValid}{DatabaseConnectorJdbcConnection}(dbObj, ...)

\S4method{dbSendQuery}{DatabaseConnectorJdbcConnection,character}(conn, statement, translate = TRUE, ...)
}
\description{
DatabaseConnectorJdbcConnection class.
}
\keyword{internal}
