% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorResult-class}
\alias{DatabaseConnectorResult-class}
\alias{dbHasCompleted,DatabaseConnectorResult-method}
\alias{dbColumnInfo,DatabaseConnectorResult-method}
\alias{dbGetStatement,DatabaseConnectorResult-method}
\alias{dbGetRowCount,DatabaseConnectorResult-method}
\alias{dbFetch,DatabaseConnectorResult-method}
\alias{dbClearResult,DatabaseConnectorResult-method}
\alias{dbGetRowsAffected,DatabaseConnectorResult-method}
\title{DatabaseConnector results class.}
\usage{
\S4method{dbHasCompleted}{DatabaseConnectorResult}(res, ...)

\S4method{dbColumnInfo}{DatabaseConnectorResult}(res, ...)

\S4method{dbGetStatement}{DatabaseConnectorResult}(res, ...)

\S4method{dbGetRowCount}{DatabaseConnectorResult}(res, ...)

\S4method{dbFetch}{DatabaseConnectorResult}(res, n = -1, ...)

\S4method{dbClearResult}{DatabaseConnectorResult}(res, ...)

\S4method{dbGetRowsAffected}{DatabaseConnectorResult}(res, ...)
}
\description{
DatabaseConnector results class.
}
\keyword{internal}
