% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{dateAdd}
\alias{dateAdd}
\title{Add an interval to a date}
\usage{
dateAdd(interval, number, date)
}
\arguments{
\item{interval}{Unit for the interval. Can be "day", "week", "month", "year".}

\item{number}{The number of units to add to the date.}

\item{date}{The date to add to.}
}
\value{
A new date.
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
dateAdd("day", 10, as.Date("2000-01-01"))

}
