% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{dateDiff}
\alias{dateDiff}
\title{Compute difference between dates}
\usage{
dateDiff(interval, date1, date2)
}
\arguments{
\item{interval}{Unit for the interval. Can be "day", "week", "month", "year".}

\item{date1}{The first date.}

\item{date2}{The second date.}
}
\value{
The numeric value of the difference.
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
dateDiff("day", as.Date("2000-01-01"), as.Date("2000-03-01"))

}
