% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{dateFromParts}
\alias{dateFromParts}
\title{Construct a date from parts}
\usage{
dateFromParts(year, month, day)
}
\arguments{
\item{year}{The calendar year.}

\item{month}{The calendar month (1 = January).}

\item{day}{The day of the month.}
}
\value{
The date.
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
dateFromParts(2000, 1, 5)

}
