% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{day}
\alias{day}
\title{Extract the day from a date}
\usage{
day(date)
}
\arguments{
\item{date}{The date.}
}
\value{
The day
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
day(as.Date("2000-02-01"))

}
