% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplyr.R
\name{db_begin.JDBCConnection}
\alias{db_begin.JDBCConnection}
\title{# Doesn't return TRUE for temporary tables}
\usage{
\method{db_begin}{JDBCConnection}(con, ...)
}
\description{
# Doesn't return TRUE for temporary tables
}

