% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\name{dbConnect,DatabaseConnectorDriver-method}
\alias{dbConnect,DatabaseConnectorDriver-method}
\title{#' Microsoft SQL Server class.
#'
#' @keywords internal
#' @export
#' @import DBI
setClass("Microsoft SQL Server",
contains = "DBIConnection"
)}
\usage{
\S4method{dbConnect}{DatabaseConnectorDriver}(drv, ...)
}
\arguments{
\item{drv}{The result of the \code{\link[=DatabaseConnectorDriver]{DatabaseConnectorDriver()}} function}

\item{...}{Other parameters. These are the same as expected by the \code{\link[=connect]{connect()}} function.}
}
\value{
Returns a DatabaseConnectorConnection object that can be used with most of the other functions in
this package.
}
\description{
Connect to a database. This function is synonymous with the \code{\link[=connect]{connect()}} function. except
a dummy driver needs to be specified
}
\details{
#' DatabaseConnectorConnection class.
#'
#' @keywords internal
#' @export
#' @import DBI
setClass("DatabaseConnectorConnection",
contains = "Microsoft SQL Server",
slots = list(
identifierQuote = "character",
stringQuote = "character", dbms = "character", uuid = "character"
)
)
#' DatabaseConnectorJdbcConnection class.
#'
#' @keywords internal
#' @export
#' @import rJava
setClass("DatabaseConnectorJdbcConnection",
contains = "DatabaseConnectorConnection",
slots = list(jConnection = "jobjRef")
)

#' DatabaseConnectorDbiConnection class.
#'
#' @keywords internal
#' @export
#' @import DBI
setClass("DatabaseConnectorDbiConnection",
contains = "DatabaseConnectorConnection",
slots = list(
dbiConnection = "DBIConnection",
server = "character"
)
)
Create a connection to a DBMS
}
\examples{
\dontrun{
conn <- dbConnect(DatabaseConnectorDriver(),
  dbms = "postgresql",
  server = "localhost/ohdsi",
  user = "joe",
  password = "secret"
)
querySql(conn, "SELECT * FROM cdm_synpuf.person;")
dbDisconnect(conn)
}

}
