% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{dbFetch,DatabaseConnectorResult-method}
\alias{dbFetch,DatabaseConnectorResult-method}
\title{Retrieve records from Kazam query}
\usage{
\S4method{dbFetch}{DatabaseConnectorResult}(res, n = -1, ...)
}
\description{
Retrieve records from Kazam query
}

