% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{dbGetQuery,DatabaseConnectorConnection,ANY-method}
\alias{dbGetQuery,DatabaseConnectorConnection,ANY-method}
\title{Send a query to DatabaseConnector}
\usage{
\S4method{dbGetQuery}{DatabaseConnectorConnection,ANY}(conn, statement, ...)
}
\description{
Send a query to DatabaseConnector
}

