% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbGetQuery.ffdf}
\alias{dbGetQuery.ffdf}
\title{Retrieve data from server as ffdf object.}
\usage{
dbGetQuery.ffdf(connection, query = "", batchSize = 1e+05)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{query}{The SQL statement to retrieve the data}

\item{batchSize}{The number of rows that will be retrieved at a time from the server. A larger
batchSize means less calls to the server so better performance, but too large a batchSize could
lead to out-of-memory errors.}
}
\value{
A ffdf object containing the data. If there are 0 rows, a regular data frame is returned instead (ffdf cannot have 0 rows)
}
\description{
This allows very large data sets to be retrieved without running out of memory.
}
\details{
Retrieves data from the database server and stores it in an ffdf object. This allows very large
data sets to be retrieved without running out of memory.
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="mysql", server="localhost",user="root",password="blah",schema="cdm_v4")
  conn <- connect(connectionDetails)
  dbGetQuery.ffdf(conn,"SELECT * FROM person")
  dbDisconnect(conn)
}
}

