% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbGetQueryPostgreSql}
\alias{dbGetQueryPostgreSql}
\title{Retrieve data from server using PostgreSQL specific commands.}
\usage{
dbGetQueryPostgreSql(connection, query = "")
}
\arguments{
\item{connection}{The connection to the database server.}

\item{query}{The SQL statement to retrieve the data}
}
\value{
A data frame containing the data retrieved from the server
}
\description{
This function is tailored to retrieve large datasets from a PostgreSQL database.
Specifically, it temporarily disables auto commit and calls \code{setFetchSize}
on the Statement object. Without these settings, all rows would be fetched
from the server, resulting in out-of-memory errors.
}
\details{
Retrieves data from the database server and stores it in a data frame.
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="postgresql", server="localhost/ohdsi",user="postgres",password="blah",schema="cdm_v4")
  conn <- connect(connectionDetails)
  dbGetQueryPostgreSql(conn,"SELECT * FROM person")
  dbDisconnect(conn)
}
}

