% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\name{dbGetStatement,DatabaseConnectorDbiResult-method}
\alias{dbGetStatement,DatabaseConnectorDbiResult-method}
\title{Get the statement associated with a result set}
\usage{
\S4method{dbGetStatement}{DatabaseConnectorDbiResult}(res, ...)
}
\arguments{
\item{res}{An object inheriting from \link[DBI:DBIResult-class]{DBIResult}.}

\item{...}{Other arguments passed on to methods.}
}
\value{
\code{dbGetStatement()} returns a string, the query used in
either \code{\link[DBI:dbSendQuery]{dbSendQuery()}} or
\code{\link[DBI:dbSendStatement]{dbSendStatement()}}.
}
\description{
Returns the statement that was passed to \code{\link[DBI:dbSendQuery]{dbSendQuery()}}
or \code{\link[DBI:dbSendStatement]{dbSendStatement()}}.

\Sexpr[results=rd,stage=render]{DBI:::methods_as_rd("dbGetStatement")}
}
\seealso{
Other DBIResult generics: 
\code{\link[DBI]{DBIResult-class}},
\code{\link[DBI]{dbBind}()},
\code{\link[DBI]{dbClearResult}()},
\code{\link[DBI]{dbColumnInfo}()},
\code{\link[DBI]{dbFetch}()},
\code{\link[DBI]{dbGetInfo}()},
\code{\link[DBI]{dbGetRowCount}()},
\code{\link[DBI]{dbGetRowsAffected}()},
\code{\link[DBI]{dbHasCompleted}()},
\code{\link[DBI]{dbIsReadOnly}()},
\code{\link[DBI]{dbIsValid}()},
\code{\link[DBI]{dbQuoteLiteral}()},
\code{\link[DBI]{dbQuoteString}()}
}
