% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbInsertTable}
\alias{dbInsertTable}
\title{Insert a table on the server}
\usage{
dbInsertTable(conn, tableName, data, dropTableIfExists = TRUE,
  append = FALSE, temp = FALSE)
}
\arguments{
\item{tableName}{The name of the table where the data should be inserted.}

\item{data}{The data frame containing the data to be inserted.}

\item{dropTableIfExists}{Drop the table first if it already exists?}

\item{append}{Append to existing table?}

\item{temp}{Should the table be created as a temp table?}

\item{connection}{The connection to the database server.}
}
\description{
This function sends the data in a data frame to a table on the server. Either a new table is
created, or the data is appended to an existing table.
}
\details{
This function sends the data in a data frame to a table on the server. Either a new table is
created, or the data is appended to an existing table.
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="mysql", server="localhost",user="root",password="blah",schema="cdm_v4")
  conn <- connect(connectionDetails)
  data <- data.frame(x = c(1,2,3), y = c("a","b","c"))
  dbInsertTable(conn,"mysql","my_table",data)
  dbDisconnect(conn)
}
}

