% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-DatabaseConnector.R
\name{dbplyr_edition.DatabaseConnectorConnection}
\alias{dbplyr_edition.DatabaseConnectorConnection}
\alias{dbplyr_edition.DatabaseConnectorJdbcConnection}
\alias{dbplyr_edition.DatabaseConnectorDbiConnection}
\title{Declare dbplyr backend edition}
\usage{
\method{dbplyr_edition}{DatabaseConnectorConnection}(con)

\method{dbplyr_edition}{DatabaseConnectorJdbcConnection}(con)

\method{dbplyr_edition}{DatabaseConnectorDbiConnection}(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Declare dbplyr backend edition
}
