% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{dbSendQuery,DatabaseConnectorConnection-method}
\alias{dbSendQuery,DatabaseConnectorConnection-method}
\title{Send a query to DatabaseConnector}
\usage{
\S4method{dbSendQuery}{DatabaseConnectorConnection}(conn, statement)
}
\description{
Send a query to DatabaseConnector
}

