% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\name{dbUnloadDriver,DatabaseConnectorDriver-method}
\alias{dbUnloadDriver,DatabaseConnectorDriver-method}
\title{Load and unload database drivers}
\usage{
\S4method{dbUnloadDriver}{DatabaseConnectorDriver}(drv, ...)
}
\arguments{
\item{drv}{an object that inherits from \code{DBIDriver} as created by
\code{dbDriver}.}

\item{...}{any other arguments are passed to the driver \code{drvName}.}
}
\value{
In the case of \code{dbDriver}, an driver object whose class extends
\code{DBIDriver}. This object may be used to create connections to the
actual DBMS engine.

In the case of \code{dbUnloadDriver}, a logical indicating whether the
operation succeeded or not.
}
\description{
These methods are deprecated, please consult the documentation of the
individual backends for the construction of driver instances.

\code{dbDriver()} is a helper method used to create an new driver object
given the name of a database or the corresponding R package. It works
through convention: all DBI-extending packages should provide an exported
object with the same name as the package. \code{dbDriver()} just looks for
this object in the right places: if you know what database you are connecting
to, you should call the function directly.

\code{dbUnloadDriver()} is not implemented for modern backends.
}
\details{
The client part of the database communication is
initialized (typically dynamically loading C code, etc.) but note that
connecting to the database engine itself needs to be done through calls to
\code{dbConnect}.
}
\seealso{
Other DBIDriver generics: 
\code{\link[DBI]{DBIDriver-class}},
\code{\link[DBI]{dbCanConnect}()},
\code{\link[DBI]{dbConnect}()},
\code{\link[DBI]{dbDataType}()},
\code{\link[DBI]{dbGetInfo}()},
\code{\link[DBI]{dbIsReadOnly}()},
\code{\link[DBI]{dbIsValid}()},
\code{\link[DBI]{dbListConnections}()}

Other DBIDriver generics: 
\code{\link[DBI]{DBIDriver-class}},
\code{\link[DBI]{dbCanConnect}()},
\code{\link[DBI]{dbConnect}()},
\code{\link[DBI]{dbDataType}()},
\code{\link[DBI]{dbGetInfo}()},
\code{\link[DBI]{dbIsReadOnly}()},
\code{\link[DBI]{dbIsValid}()},
\code{\link[DBI]{dbListConnections}()}
}
