% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from the server}
\usage{
disconnect(connection)
}
\arguments{
\item{connection}{The connection to the database server created using either
\code{\link[=connect]{connect()}} or \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Close the connection to the server.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(
  dbms = "postgresql",
  server = "localhost",
  user = "root",
  password = "blah"
)
conn <- connect(connectionDetails)
count <- querySql(conn, "SELECT COUNT(*) FROM person")
disconnect(conn)
}
}
