% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{eoMonth}
\alias{eoMonth}
\title{Return the end of the month}
\usage{
eoMonth(date)
}
\arguments{
\item{date}{A date in the month for which we need the end.}
}
\value{
The date of the last day of the month.
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
eoMonth(as.Date("2000-02-01"))

}
