% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql.R
\name{getAvailableJavaHeapSpace}
\alias{getAvailableJavaHeapSpace}
\title{Get available Java heap space}
\usage{
getAvailableJavaHeapSpace()
}
\value{
The Java heap space (in bytes).
}
\description{
For debugging purposes: get the available Java heap space.
}
