% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\name{inDatabaseSchema}
\alias{inDatabaseSchema}
\title{Refer to a table in a database schema}
\usage{
inDatabaseSchema(databaseSchema, table)
}
\arguments{
\item{databaseSchema}{The name of the database schema. For example, for SQL Server this can 
be 'cdm.dbo'.}

\item{table}{The name of the table in the database schema.}
}
\value{
An object representing the table and database schema.
}
\description{
Can be used with \code{dplyr::tbl()} to indicate a table in a specific database schema.
}
