% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverDownload.R
\name{installJdbcDriver}
\alias{installJdbcDriver}
\title{Install JDBC driver in package}
\usage{
installJdbcDriver(dbms)
}
\arguments{
\item{dbms}{The type of DBMS driver to install. Valid values are
\itemize{
  \item {"oracle" for Oracle}
  \item {"postgresql" for PostgreSQL}
  \item {"redshift" for Amazon Redshift}
  \item {"sql server" for Microsoft SQL Server}
  \item {"pdw" for Microsoft Parallel Data Warehouse (PDW)}
}}
}
\description{
Install JDBC driver in package
}
\details{
Installs the appropriate JDBC driver from the internet.
}

