% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql.R
\name{isSqlReservedWord}
\alias{isSqlReservedWord}
\title{Test a character vector of SQL names for SQL reserved words}
\usage{
isSqlReservedWord(sqlNames, warn = FALSE)
}
\arguments{
\item{sqlNames}{A character vector containing table or field names to check.}

\item{warn}{(logical) Should a warn be thrown if invalid SQL names are found?}
}
\value{
A logical vector with length equal to sqlNames that is TRUE for each name that is reserved and FALSE otherwise
}
\description{
This function checks a character vector against a predefined list of reserved SQL words.
}
