% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnector.R
\name{jdbcDrivers}
\alias{jdbcDrivers}
\title{How to download and use JDBC drivers for the various data platforms.}
\description{
Below are instructions for downloading JDBC drivers for the various data platforms. Once downloaded
use the \code{pathToDriver} argument in the \code{\link{connect}} or \code{\link{createConnectionDetails}}
functions to point to the driver. Alternatively, you can set the 'DATABASECONNECTOR_JAR_FOLDER' environmental
variable, for example in your .Renviron file (recommended).
}
\section{SQL Server, Oracle, PostgreSQL, PDW, Spark, RedShift}{
 Use the \code{\link{downloadJdbcDrivers}} function to download these drivers
from the OHDSI GitHub pages.
}

\section{Netezza}{
 Read the instructions
\href{https://www.ibm.com/docs/en/SSULQD_7.2.1/com.ibm.nz.datacon.doc/t_datacon_setup_JDBC.html}{here}
on how to obtain the Netezza JDBC driver.
}

\section{BigQuery}{
 Go to \href{https://cloud.google.com/bigquery/docs/reference/odbc-jdbc-drivers}{Google's site} and
download the latest JDBC driver. Unzip the file, and locate the appropriate jar files.
}

\section{Impala}{
 Go to
\href{https://www.cloudera.com/downloads/connectors/impala/jdbc/2-5-5.html}{Cloudera's site}, pick
your OS version, and click "GET IT NOW!'. Register, and you should be able to download the driver.
}

\section{SQLite}{
 For SQLite we actually don't use a JDBC driver. Instead, we use the RSQLite package, which can be installed
using \code{install.packages("RSQLite")}.
}

\section{Spark}{
 Go to \href{https://databricks.com/spark/jdbc-drivers-download}{Databricks JDBC site} and 
download the latest JDBC driver (license may be required).
}

