% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql.R
\name{lowLevelExecuteSql}
\alias{lowLevelExecuteSql}
\title{Execute SQL code}
\usage{
lowLevelExecuteSql(connection, sql)
}
\arguments{
\item{connection}{The connection to the database server created using either
\code{\link[=connect]{connect()}} or \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{sql}{The SQL to be executed}
}
\description{
This function executes a single SQL statement.
}
