% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalRsqlite.R
\name{lowLevelQuerySql.sqlite}
\alias{lowLevelQuerySql.sqlite}
\title{Low level function for retrieving data to a local SQLite database}
\usage{
lowLevelQuerySql.sqlite(
  connection,
  query,
  sqliteConnection,
  sqliteTableName,
  datesAsString = FALSE
)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{query}{The SQL statement to retrieve the data}

\item{sqliteConnection}{An open connection to a SQLite database, for example as created using \code{RSQLite::dbConnect()}.}

\item{sqliteTableName}{The name of the table in the local SQLite database where the results of the query will be stored.}

\item{datesAsString}{Should dates be imported as character vectors, our should they be converted
to R's date format?}
}
\value{
Invisibly returns the sqliteConnection. The SQLite database will have a table added with the query results.
}
\description{
This is the equivalent of the \code{\link{querySql.sqlite}} function, except no error report is
written when an error occurs.
}
\details{
Retrieves data from the database server and stores it in a local SQLite database This allows very large
data sets to be retrieved without running out of memory. Null values in the database are converted
to NA values in R. If a table with the same name already exists in the local SQLite database it is replaced.
}
