% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{month}
\alias{month}
\title{Extract the month from a date}
\usage{
month(date)
}
\arguments{
\item{date}{The date.}
}
\value{
The month
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
month(as.Date("2000-02-01"))

}
