% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{querySql.ffdf}
\alias{querySql.ffdf}
\title{Send SQL query}
\usage{
querySql.ffdf(connection, sql)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{sql}{The SQL to be send.}
}
\value{
An ffdf object.
}
\description{
This function sends SQL to the server, and returns the results in an ffdf object.
}
\details{
This function sends the SQL to the server and retrieves the results. If an error occurs during
SQL execution, this error is written to a file to facilitate debugging.
}
\examples{
\dontrun{
  library(ffbase)
  connectionDetails <- createConnectionDetails(dbms="mysql",
                                               server="localhost",
                                               user="root",
                                               password="blah",
                                               schema="cdm_v4")#'   conn <- connect(connectionDetails)
  count <- querySql.ffdf(conn,"SELECT COUNT(*) FROM person")
  dbDisconnect(conn)
}
}

