% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andromeda.R
\name{renderTranslateQuerySqlToAndromeda}
\alias{renderTranslateQuerySqlToAndromeda}
\title{Render, translate, and query to local Andromeda}
\usage{
renderTranslateQuerySqlToAndromeda(
  connection,
  sql,
  andromeda,
  andromedaTableName,
  errorReportFile = file.path(getOption("errorReportFileLocation", getwd()),
    "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  appendToTable = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE),
  ...
)
}
\arguments{
\item{connection}{The connection to the database server created using either
\code{\link[=connect]{connect()}} or \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{sql}{The SQL to be send.}

\item{andromeda}{An open Andromeda object, for example as created
using \code{\link[Andromeda:andromeda]{Andromeda::andromeda()}}.}

\item{andromedaTableName}{The name of the table in the local Andromeda object where the
results of the query will be stored.}

\item{errorReportFile}{The file where an error report will be written if an error occurs.
Defaults to 'errorReportSql.txt' in the directory given by
\code{getOption("errorReportFileLocation")}, or the current working
directory if that option is unset.}

\item{snakeCaseToCamelCase}{If true, field names are assumed to use snake_case, and are converted to camelCase.}

\item{appendToTable}{If FALSE, any existing table in the Andromeda with the same name will be
replaced with the new data. If TRUE, data will be appended to an existing
table, assuming it has the exact same structure.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{integerAsNumeric}{Logical: should 32-bit integers be converted to numeric (double) values? If FALSE
32-bit integers will be represented using R's native \code{Integer} class.}

\item{integer64AsNumeric}{Logical: should 64-bit integers be converted to numeric (double) values? If FALSE
64-bit integers will be represented using \code{bit64::integer64}.}

\item{...}{Parameters that will be used to render the SQL.}
}
\value{
Invisibly returns the andromeda. The Andromeda object will have a table added with the query
results.
}
\description{
This function renders, and translates SQL, sends it to the server, and returns the results as an
\code{Andromeda} object
}
\details{
This function calls the \code{render} and \code{translate} functions in the \code{SqlRender} package
before calling \code{\link[=querySqlToAndromeda]{querySqlToAndromeda()}}.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(
  dbms = "postgresql",
  server = "localhost",
  user = "root",
  password = "blah",
  schema = "cdm_v4"
)
conn <- connect(connectionDetails)
renderTranslatequerySqlToAndromeda(conn,
  sql = "SELECT * FROM @schema.person",
  schema = "cdm_synpuf",
  andromeda = andromeda,
  andromedaTableName = "foo"
)
disconnect(conn)

andromeda$foo
}
}
