% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{setPathToDll}
\alias{setPathToDll}
\title{Set the path to the authentication DLL}
\usage{
setPathToDll(path = getwd())
}
\arguments{
\item{path}{The absolute path where the DLL can be found.}
}
\description{
Set the path to the authentication DLL
}
\details{
Set the path to the authentication DLL
}
