% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplyr.R
\name{sql_subquery.JDBCConnection}
\alias{sql_subquery.JDBCConnection}
\title{#' @export
sql_escape_ident.JDBCConnection <- function(con, x) {
  sql_quote(x, '"')
}}
\usage{
\method{sql_subquery}{JDBCConnection}(con, from, name = unique_name(), ...)
}
\description{
#' @export
sql_escape_ident.JDBCConnection <- function(con, x) {
  sql_quote(x, '"')
}
}

