% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbiDateFunctions.R
\name{year}
\alias{year}
\title{Extract the year from a date}
\usage{
year(date)
}
\arguments{
\item{date}{The date.}
}
\value{
The year
}
\description{
This function is provided primarily to be used together with \code{dbplyr} when querying
a database. It will also work in \code{dplyr} against data frames.
}
\examples{
year(as.Date("2000-02-01"))

}
