% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnector.R
\docType{package}
\name{DatabaseConnector-package}
\alias{DatabaseConnector}
\alias{DatabaseConnector-package}
\title{DatabaseConnector: Connecting to Various Database Platforms}
\description{
An R 'DataBase Interface' ('DBI') compatible interface to various database platforms ('PostgreSQL', 'Oracle', 'Microsoft SQL Server', 'Amazon Redshift', 'Microsoft Parallel Database Warehouse', 'IBM Netezza', 'Apache Impala', 'Google BigQuery', 'Snowflake', 'Spark', and 'SQLite'). Also includes support for fetching data as 'Andromeda' objects. Uses either 'Java Database Connectivity' ('JDBC') or other 'DBI' drivers to connect to databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DatabaseConnector/}
  \item \url{https://github.com/OHDSI/DatabaseConnector}
  \item Report bugs at \url{https://github.com/OHDSI/DatabaseConnector/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
  \item Microsoft Inc. (SQL Server JDBC driver) [copyright holder]
  \item PostgreSQL Global Development Group (PostgreSQL JDBC driver) [copyright holder]
  \item Oracle Inc. (Oracle JDBC driver) [copyright holder]
  \item Amazon Inc. (RedShift JDBC driver) [copyright holder]
}

}
\keyword{internal}
