% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorDbiConnection-class}
\alias{DatabaseConnectorDbiConnection-class}
\alias{dbIsValid,DatabaseConnectorDbiConnection-method}
\alias{dbSendQuery,DatabaseConnectorDbiConnection,character-method}
\title{DatabaseConnectorDbiConnection class.}
\usage{
\S4method{dbIsValid}{DatabaseConnectorDbiConnection}(dbObj, ...)

\S4method{dbSendQuery}{DatabaseConnectorDbiConnection,character}(conn, statement, translate = TRUE, ...)
}
\description{
DatabaseConnectorDbiConnection class.
}
\keyword{internal}
