% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorJdbcConnection-class}
\alias{DatabaseConnectorJdbcConnection-class}
\title{DatabaseConnectorJdbcConnection class.}
\description{
DatabaseConnectorJdbcConnection class.
}
\keyword{internal}
