% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{class}
\name{DatabaseConnectorJdbcResult-class}
\alias{DatabaseConnectorJdbcResult-class}
\title{DatabaseConnector JDBC results class.}
\description{
DatabaseConnector JDBC results class.
}
\keyword{internal}
