% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-DatabaseConnector.R
\name{dbplyr_edition.DatabaseConnectorConnection}
\alias{dbplyr_edition.DatabaseConnectorConnection}
\title{#' @export sql_values_subquery
#' @export
sql_values_subquery <- function(con, df, types, lvl = 0, ...){
check_dots_used()
UseMethod("sql_values_subquery")
}}
\usage{
\method{dbplyr_edition}{DatabaseConnectorConnection}(con)
}
\description{
#' @export sql_values_subquery.DatabaseConnectorJdbcConnection
#' @export
sql_values_subquery.DatabaseConnectorJdbcConnection <- function(con, df, types, lvl = 0, ...) {
if(dbms(con) == "redshift"){
return(utils::getFromNamespace("sql_values_subquery.RedshiftConnection", "dbplyr")(con = con, df = df, types = types, lvl = lvl, ...))
}
else{
return(utils::getFromNamespace("sql_values_subquery.DBIConnection", "dbplyr")(con = con, df = df, types = types, lvl = lvl, ...))
}
}
}
