% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from the server}
\usage{
disconnect(connection)
}
\arguments{
\item{connection}{The connection to the database server.}
}
\description{
This function sends SQL to the server, and returns the results in an ffdf object.
}
\examples{
\dontrun{
library(ffbase)
connectionDetails <- createConnectionDetails(dbms = "mysql",
                                             server = "localhost",
                                             user = "root",
                                             password = "blah",
                                             schema = "cdm_v4")
conn <- connect(connectionDetails)
count <- querySql.ffdf(conn, "SELECT COUNT(*) FROM person")
disconnect(conn)
}
}
