% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\name{inDatabaseSchema}
\alias{inDatabaseSchema}
\title{Refer to a table in a database schema}
\usage{
inDatabaseSchema(databaseSchema, table)
}
\arguments{
\item{databaseSchema}{}

\item{table}{}
}
\value{
An object representing the table and database schema.
}
\description{
Can be used with \code{dplyr::tbl()} to indicate a table in a specific database schema.
}
