% Generated by roxygen2 (4.0.2): do not edit by hand
\name{querySql}
\alias{querySql}
\title{Send SQL query}
\usage{
querySql(conn, sql)
}
\arguments{
\item{sql}{The SQL to be send.}

\item{connection}{The connection to the database server.}
}
\description{
This function sends SQL to the server, and returns the results.
}
\details{
This function sends the SQL to the server and retrieves the results. If an error occurs during
SQL execution, this error is written to a file to facilitate debugging.
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="mysql", server="localhost",user="root",password="blah",schema="cdm_v4")
  conn <- connect(connectionDetails)
  count <- querySql(conn,"SELECT COUNT(*) FROM person")
  dbDisconnect(conn)
}
}

