% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-DatabaseConnector.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sql_query_select}
\alias{sql_escape_logical}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dbplyr}{\code{\link[dbplyr:db-quote]{sql_escape_logical}}, \code{\link[dbplyr:db-sql]{sql_query_select}}}
}}

